Better FlatZinc error messages
When a FlatZinc::Error is constructed, the annotations can be passed
to the constructor for additional information. Currently, this looks
through the annotations, and if any 'mzn_constraint_name' annotation
is found, this is added to the error string.

Thus, an error that was previously reported as
    Error: Gecode: Float::linear: Number out of limits
will now (given that the constraint the error originates from has been
annotated in the MiniZinc file) be reported as
    Error: Gecode: Float::linear: Number out of limits in constraint "my constraint name"