Make make check work when no dynamic load path exists
When running make check, the test binary tries to load the shared
libraries it depends on. However, if the user has no dynamic load path
set (LD_LIBRARY_PATH on *nix, DYLD_LIBRARY_PATH on mac, no equivalent
on Windows), then runing the binary will fail since the librareis to
load are not installed on the system yet.

This commit adds a new configuration element, that configures a prefix
to use that sets the correct dynamic load path.

Note that Windows has no concept of a dynamic load path, so for
Windows nothing is set.