Avoid reserved identifier names
Identifiers with double underscores and identifiers starting with an
underscore followed by a captial letter are reserved, and using them
is undefined behaviour.  See more at
https://en.cppreference.com/w/cpp/language/identifiers

This change removes all the remaining cases in Gecode proper that
declares names with double underscores and those starting with a
single underscroe followed by a capital letter.

This fixes the remaining issues from issue #63