Fix path compression in disjoint sets for sorted
The OfflineMin data structure used for the sorted constraint is a
union-find disjoint sets data structure. The find_pc method was
supposed to do path compression by storing all the nodes seen during
traversal of the tree, and then later setting all the intermediaries
to the root.

The code did not work as intended, it never did any path compression
at all.

No change in test behaviour for sorted (still works, and no changee in
time).