cmake: Remove unused checks.
None of the header checks were used for that and can be removed.

One was used to check for `sys/time.h` and then enable `gettimeofday()`
so we can instead just check for the `gettimeofday()` function instead.

Remove unused check for size of `int`.

Remove unused check for `getpagesize()`.

Remove `GECODE_HAS_UNISTD_H` from the config header as it isn't
used or set by anything (any longer).